优化缓存管理与节点逻辑实现
1. **缓存管理优化**：
   - 在 `CacheMgr` 中增加对 `nid` 的支持，确保缓存键的唯一性。
   - 新增 `buildCache4GenerateKey` 方法，简化缓存键生成逻辑。
   - 在多个节点（如 `TaskNode`、`IterRun` 等）中调整缓存键生成逻辑，统一使用 `buildCache4GenerateKey`。

2. **节点逻辑改进**：
   - 修复 `attach_node_input` 和 `attach_node_output` 中的缓存键生成问题。
   - 优化 `IterRun` 节点的迭代逻辑，确保子节点与附属节点的正确连接与执行。
   - 调整 `TaskNode` 的 `addPreNode` 方法，确保前置节点的正确性。

3. **代码清理与一致性**：
   - 删除无用的测试文件 `testdb4cache.py`。
   - 调整部分代码格式与注释，提升可读性。
   - 修复部分拼写错误与不一致的代码风格。

4. **其他改动**：
   - 在 `VFNodeInterface` 中调整 `VFNodeSize` 的类型为 `float`，以支持更精确的节点尺寸。
   - 更新部分节点的 `getCacheKey` 方法逻辑，确保缓存的正确性与性能优化。

+++ backend/app/schemas/VFNodeInterface.py
@@ -146,8 +146,8 @@ class VFNodePadding(BaseModel):
 
 
 class VFNodeSize(BaseModel):
-    Width: int
-    Height: int
+    Width: float
+    Height: float
 
 
 class VFNodeNesting(BaseModel):
